refactor: standalone `Not Bundle` plugin
