+++ .gitignore
@@ -109,4 +109,5 @@ package-lock.json
 # pnpm-lock.yaml
 yarn.lock
 
+/plugin
 /types
