+++ test/utils.test.ts
@@ -1,20 +1,12 @@
-import fs from 'node:fs'
-import path from 'node:path'
 import { builtinModules } from 'node:module'
 import { ExternalOption } from 'rollup'
-import {
-  type InlineConfig,
-  build,
-} from 'vite'
+import type { InlineConfig } from 'vite'
 import {
   describe,
   expect,
   it,
 } from 'vitest'
-import {
-  withExternalBuiltins,
-  external_node_modules,
-} from '../src/utils'
+import { withExternalBuiltins } from '..'
 
 const builtins: any[] = builtinModules.filter(e => !e.startsWith('_')); builtins.push('electron', ...builtins.map(m => `node:${m}`))
 const getConfig = (external: ExternalOption): InlineConfig => ({ build: { rollupOptions: { external } } })
@@ -36,28 +28,5 @@ describe('src/config', () => {
 
     const external_fun = withExternalBuiltins(getConfig(external_function))!.build!.rollupOptions!.external
     expect((external_fun as (source: string) => boolean)('electron')).true
-  })
-
-  it('external_node_modules', async () => {
-    await build({
-      configFile: false,
-      root: __dirname,
-      build: {
-        lib: {
-          entry: 'fixtures/external-main.ts',
-          formats: ['cjs'],
-          fileName: () => 'external-main.js',
-        },
-        minify: false,
-      },
-      plugins: [
-        external_node_modules(),
-      ],
-    })
-
-    const distMain = fs.readFileSync(path.join(__dirname, '__snapshots__/external-main.js'), 'utf-8')
-    const snapMain = fs.readFileSync(path.join(__dirname, 'dist/external-main.js'), 'utf-8')
-
-    expect(distMain).equal(snapMain)
   })
 })
